Factor out sign up and log in
