+++ src/App.css
@@ -1,7 +1,7 @@
 .container {
   height: 100vh;
   position: relative;
-  top: 65px;
+  top: 80px; /* offset nav height */
   font-family: Georgia, 'Times New Roman', Times, serif;
   display: flex;
   flex-direction: column;
