--- src/components/authForm/AuthForm.module.css
@@ -1,66 +1,0 @@
-.container {
-  width: 300px;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-}
-
-form > * {
-  margin-top: 5px;
-  margin-bottom: 5px;
-  border-radius: 15px;
-  font-weight: bold;
-}
-
-.btn {
-  border-radius: 20px;
-  background-color: transparent;
-  border: 2px solid var(--primary-color-site);
-}
-
-.btn:hover {
-  cursor: pointer;
-  border-color: black;
-  background-color: var(--primary-color-site);
-}
-
-.toggle,
-.submit {
-  background-color: pink;
-}
-
-.toggle {
-  border-color: var(--secondary-color-site);
-  width: 100%;
-  padding: 12px 10px;
-  margin: 8px 0;
-  font-weight: bold;
-}
-
-/* Potentially whack button hover flair */
-.toggle:hover {
-  color: pink;
-  border-color: pink;
-  background-color: var(--secondary-color-site);
-}
-
-.submit:hover {
-  color: pink;
-  border-color: pink;
-}
-
-.btn,
-input {
-  height: 2.5rem;
-}
-
-.formGroup {
-  display: flex;
-  flex-direction: column;
-}
-
-.title {
-  text-align: center;
-}
